edit + delete msg api request + ws implement for edit and delete msg in a conv

+++ backend/src/controllers/messageController.js
@@ -64,17 +64,25 @@ exports.updateMessage = async (req, res) => {
 
 exports.deleteMessage = async (req, res) => {
     const { id } = req.params;
+    const userId = req.user._id;
 
     try {
+        const message = await messageModel.findById(id);
+
+        if (!message) {
+            return res.status(404).json({ message: "Message not found." });
+        }
+
+        if (message.sender.toString() !== userId.toString()) {
+            return res.status(403).json({ message: "Unauthorized to delete this message." });
+        }
+
         await messageModel.findByIdAndDelete(id);
-
         global.io.emit('messageDeleted', { messageId: id });
-
         res.status(200).json({ message: 'Message deleted successfully' });
-    }
-
-    catch (error) {
+    } catch (error) {
         res.status(500).json({ message: error.message });
     }
-}
+};
+
